requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
duckduckgo-search
pypdf
python-dotenv